+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.9.0(2021-07-12)
+
+- Update dependencies
+- Use `ContextVar` to find correct components in `use_component`
+- Use `ContextVar` to find correct api instance in `use_api`
+- Fix `cli.urls_ts` to give better type information
+
 ## v0.8.0(2020-12-07)
 
 - Change target python version>=3.9